Incorporate feedback from Paul's review: Add documentation and remove references to old config, mainly
