+++ portal/views/auth.py
@@ -1,7 +1,6 @@
 """Auth related view functions"""
 from __future__ import unicode_literals  # isort:skip
 
-from abc import ABCMeta, abstractmethod
 import base64
 from datetime import datetime
 import hashlib
@@ -42,6 +41,12 @@ from ..models.client import client_event_dispatch, validate_origin
 from ..models.coredata import Coredata
 from ..models.encounter import finish_encounter
 from ..models.login import login_user
+from ..models.flaskdanceprovider import (
+    FacebookFlaskDanceProvider,
+    FlaskProviderUserInfo,
+    GoogleFlaskDanceProvider,
+    MockFlaskDanceProvider,
+)
 from ..models.role import ROLE
 from ..models.user import (
     User,
@@ -63,168 +68,6 @@ facebook_blueprint = make_facebook_blueprint(
 )
 
 
-class FlaskDanceProvider:
-    """base class for flask dance providers
-
-    When a new provider is added to the protal's consumer oauth flow
-    a descendent of this class needs to be created to get the user's
-    information from the provider after a successful auth
-    """
-    __metaclass__ = ABCMeta
-
-    def __init__(self, blueprint, token):
-        self.blueprint = blueprint
-        self.token = token
-
-    @property
-    def name(self):
-        return self.blueprint.name
-
-    @abstractmethod
-    def get_user_info(self):
-        """gets user info from the provider
-
-        This function must be overriden in descendant classes
-        to return an instance of FlaskProviderUserInfo that is
-        filled with user information fetched from the provider
-        """
-        pass
-
-
-class FacebookFlaskDanceProvider(FlaskDanceProvider):
-    """fetches user info from Facebook after successfull auth
-
-    After the user successfully authenticates with Facebook
-    this class fetches the user's info from Facebook and packages
-    it in FlaskDanceProviderUserInfo
-    """
-
-    def get_user_info(self):
-        """gets user info from Facebook
-
-        After the user successfully authenticates with Facebook
-        control enters this function which gets the user's info from
-        Facebook and returns an instance of FlaskDanceProviderUserInfo
-        """
-        resp = self.blueprint.session.get(
-            '/me',
-            params={
-                'fields':
-                    'id,email,birthday,first_name,last_name,gender,picture'})
-        if not resp.ok:
-            current_app.logger.debug('Failed to fetch user info from Facebook')
-            return False
-
-        facebook_info = resp.json()
-
-        user_info = FlaskProviderUserInfo()
-        user_info.id = facebook_info['id']
-        user_info.first_name = facebook_info['first_name']
-        user_info.last_name = facebook_info['last_name']
-        user_info.email = facebook_info['email']
-        user_info.gender = facebook_info['gender']
-        user_info.image_url = facebook_info['picture']['data']['url']
-        user_info.birthdate = facebook_info['birthday']
-
-        return user_info
-
-
-class GoogleFlaskDanceProvider(FlaskDanceProvider):
-    """fetches user info from Google after successfull auth
-
-    After the user successfully authenticates with Google
-    this class fetches the user's info from Google and packages it
-    in FlaskDanceProviderUserInfo
-    """
-
-    def get_user_info(self):
-        """gets user info from Google
-
-        After the user successfully authenticates with Google
-        control enters this function which gets the user's info
-        from Google and returns an instance of FlaskDanceProviderUserInfo
-        """
-        resp = self.blueprint.session.get('/oauth2/v2/userinfo')
-        if not resp.ok:
-            current_app.logger.debug('Failed to fetch user info from Google')
-            return False
-
-        google_info = resp.json()
-
-        user_info = FlaskProviderUserInfo()
-        user_info.id = google_info['id']
-        user_info.first_name = google_info['given_name']
-        user_info.last_name = google_info['family_name']
-        user_info.email = google_info['email']
-        user_info.image_url = google_info['picture']
-
-        # Gender may not be available
-        if 'gender' in google_info:
-            user_info.gender = google_info['gender']
-
-        # Birthday may not be available
-        if 'birthday' in google_info:
-            user_info.birthdate = google_info['birthday']
-
-        return user_info
-
-
-class MockFlaskDanceProvider(FlaskDanceProvider):
-    """creates user info from test data to validate auth logic
-
-    This class should only be used during testing.
-    It simply returns the data passed into its constructor in
-    get_user_info. This effectively mocks out the get_user_info
-    request that's normally sent to a provider after successful oauth
-    in non-test environments.
-    """
-
-    def __init__(self, provider_name, token, user_info, fail_to_get_user_info):
-        blueprint = type(str('MockBlueprint'), (object,), {})
-        blueprint.name = provider_name
-        super(MockFlaskDanceProvider, self).__init__(
-                blueprint,
-                token)
-
-        self.user_info = user_info
-        self.fail_to_get_user_info = fail_to_get_user_info
-
-    def get_user_info(self):
-        """return the user info passed into the constructor
-
-        This effectively mocks out the get_user_info
-        request that's normally sent to a provider after successful oauth
-        in non-test environments.
-        """
-        if self.fail_to_get_user_info:
-            current_app.logger.debug(
-                'MockFlaskDanceProvider failed to get user info'
-            )
-            return False
-
-        return self.user_info
-
-
-class FlaskProviderUserInfo(object):
-    """a common format for user info fetched from providers
-
-    Each provider packages user info a litle differently.
-    Google, for example, uses "given_name" and the key for the user's
-    first name, and Facebook uses "first_name". To make it easier for
-    our code to parse responses in a common function this class provides a
-    common format to store the results from each provider.
-    """
-
-    def __init__(self):
-        self.id = None
-        self.first_name = None
-        self.last_name = None
-        self.email = None
-        self.birthdate = None
-        self.gender = None
-        self.image_url = None
-
-
 @auth.route('/test/oauth')
 def oauth_test_backdoor():
     """unit test backdoor
@@ -234,7 +77,6 @@ def oauth_test_backdoor():
     this API will likely be used.
     """
     def login_test_user_using_session(user_id):
-        assert int(user_id) < 10  # allowed for test users only!
         session['id'] = user_id
         user = current_user()
         login_user(user, 'password_authenticated')
@@ -327,7 +169,7 @@ def login_user_with_provider(request, provider):
     # Use the auth token to get user info from the provider
     user_info = provider.get_user_info()
     if not user_info:
-        current_app.logger.debug(
+        current_app.logger.error(
             'Failed to get user info at %s',
             provider.name
         )
@@ -447,7 +289,7 @@ def provider_oauth_error(
     """
     reload_count = session.get('force_reload_count', 0)
     if reload_count > 2:
-        current_app.logger.warn(
+        current_app.logger.error(
             "Failed 3 attempts: OAuth error from {name}! "
             "error={error} description={description} uri={uri}"
         ).format(
@@ -494,7 +336,7 @@ def deauthorized():
     sig = base64_url_decode(encoded_sig)
     data = base64_url_decode(payload)
 
-    secret = current_app.config['FB_CONSUMER_SECRET']
+    secret = current_app.config['FACEBOOK_OAUTH_CLIENT_SECRET']
     expected_sig = hmac.new(
         secret, msg=payload, digestmod=hashlib.sha256).digest()
     if expected_sig != sig:
